Add unittests under common structure
